Tracing and reform
