+++ src/learnatvcs.rs/Cargo.toml
@@ -23,4 +23,6 @@ serde = { version = "^1.0.147", features = ["derive"] }
 serde_json = "1.0.88"
 thiserror = "^1.0.37"
 tokio = { version = "^1.21.2", features = ["full"] }
+tracing = "0.1.37"
+tracing-subscriber = "0.3.16"
 # tl = { version = "0.7.7", features = ["simd"] }
